Made the Jython object factory reusable by other Jython connections.
