+++ org.eclipse.scanning.points/META-INF/MANIFEST.MF
@@ -14,7 +14,8 @@ Require-Bundle: org.eclipse.scanning.api;bundle-version="1.0.0",
  com.fasterxml.jackson.core.jackson-core;bundle-version="2.2.0",
  com.fasterxml.jackson.core.jackson-databind;bundle-version="2.2.0",
  org.apache.commons.math3
-Export-Package: org.eclipse.scanning.points,
+Export-Package: org.eclipse.scanning.jython,
+ org.eclipse.scanning.points,
  org.eclipse.scanning.points.classregistry,
  org.eclipse.scanning.points.mutators,
  org.eclipse.scanning.points.serialization,
