+++ org.eclipse.scanning.points/src/org/eclipse/scanning/points/SpiralIterator.java
@@ -17,7 +17,7 @@ import java.util.Iterator;
 import org.eclipse.scanning.api.points.IPosition;
 import org.eclipse.scanning.api.points.Point;
 import org.eclipse.scanning.api.points.models.SpiralModel;
-import org.eclipse.scanning.points.ScanPointGeneratorFactory.JythonObjectFactory;
+import org.eclipse.scanning.jython.JythonObjectFactory;
 import org.python.core.PyList;
 
 class SpiralIterator extends AbstractScanPointIterator {
