+++ org.eclipse.scanning.test/src/org/eclipse/scanning/test/points/ExtensionPointPluginTest.java
@@ -11,13 +11,13 @@
  *******************************************************************************/
 package org.eclipse.scanning.test.points;
 
+import static org.junit.Assert.assertNotNull;
+
 import org.eclipse.scanning.api.points.IPointGenerator;
 import org.eclipse.scanning.api.points.IPointGeneratorService;
-import org.eclipse.scanning.api.points.Point;
 import org.eclipse.scanning.points.PointGeneratorService;
 import org.junit.Before;
 import org.junit.Test;
-import static org.junit.Assert.assertNotNull;
 
 // Run as plugin test
 public class ExtensionPointPluginTest {
