+++ org.eclipse.scanning.test/src/org/eclipse/scanning/test/points/PointServiceTest.java
@@ -13,7 +13,6 @@ package org.eclipse.scanning.test.points;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertSame;
 
 import org.eclipse.dawnsci.analysis.api.roi.IRectangularROI;
 import org.eclipse.dawnsci.analysis.dataset.roi.RectangularROI;
