+++ org.eclipse.scanning.test/src/org/eclipse/scanning/test/points/RasterTest.java
@@ -11,7 +11,7 @@
  *******************************************************************************/
 package org.eclipse.scanning.test.points;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 import java.util.Iterator;
 import java.util.List;
