+++ org.eclipse.scanning.test/src/org/eclipse/scanning/test/points/ScanPointGeneratorFactoryTest.java
@@ -23,9 +23,9 @@ import org.eclipse.scanning.api.points.IPosition;
 import org.eclipse.scanning.api.points.Point;
 import org.eclipse.scanning.api.points.Scalar;
 import org.eclipse.scanning.api.points.models.ScanRegion;
+import org.eclipse.scanning.jython.JythonObjectFactory;
 import org.eclipse.scanning.points.CompoundSpgIterator;
 import org.eclipse.scanning.points.ScanPointGeneratorFactory;
-import org.eclipse.scanning.points.ScanPointGeneratorFactory.JythonObjectFactory;
 import org.junit.Test;
 import org.python.core.PyDictionary;
 import org.python.core.PyList;
